Move quick editor to a seperate file
